+++ src/pages/login.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import { Link, Redirect } from 'react-router-dom';
+import { connect } from 'react-redux';
 import logo from '../trivia.png';
 import '../App.css';
+import { getLogin } from '../actions';
 
 class Login extends React.Component {
   constructor() {
@@ -19,14 +21,18 @@ class Login extends React.Component {
   }
 
   async handlePlayButton() {
+    const { name, email } = this.state;
+    const { dispatchUserInfo } = this.props;
     fetch('https://opentdb.com/api_token.php?command=request')
       .then((res) => {
         res.json()
           .then((json) => {
             localStorage.setItem('token', json.token);
             this.setState({ play: true });
+            dispatchUserInfo(name, email);
           });
       });
+
   }
 
   checkForm() {
@@ -90,4 +96,8 @@ class Login extends React.Component {
   }
 }
 
-export default Login;
+const mapDispatchToProps = (dispatch) => ({
+  dispatchUserInfo: (name, email) => dispatch(getLogin(name, email)),
+});
+
+export default connect(null, mapDispatchToProps)(Login);
